--- a/edwards25519/cmove_decl.go
+++ b/edwards25519/cmove_decl.go
@@ -1,0 +1,14 @@
+// Copyright 2013 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// +build amd64
+
+package edwards25519
+
+// Replace (f,g) with (g,g) if b == 1;
+// replace (f,g) with (f,g) if b == 0.
+//
+// Preconditions: b in {0,1}.
+//go:noescape
+func FeCMove(f, g *FieldElement, b int32)